Move tests from NotificationsLogicTest to NotificationsDbTest
